+++ services/madoc-ts/src/frontend/shared/viewers/universal-viewer.lazy.tsx
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export const UniversalViewer = React.lazy(() => import('./universal-viewer'));